requirements.txt
flask
flask_restful
Flask-SQLAlchemy
flask-migrate
python-dotenv